+++ rocprim/include/rocprim/thread/thread_operators.hpp
@@ -1,7 +1,7 @@
 /******************************************************************************
  * Copyright (c) 2010-2011, Duane Merrill.  All rights reserved.
  * Copyright (c) 2011-2018, NVIDIA CORPORATION.  All rights reserved.
- * Modifications Copyright (c) 2017-2025, Advanced Micro Devices, Inc.  All rights reserved.
+ * Modifications Copyright (c) 2017-2026, Advanced Micro Devices, Inc.  All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -39,14 +39,8 @@
 
 BEGIN_ROCPRIM_NAMESPACE
 
-/// \defgroup thread_operators Thread Operators
-/// \ingroup threadmodule
-
-/// \addtogroup thread_operators
-/// @{
-
 /// \brief Functor that tests for equality.
-struct equality
+struct [[deprecated("Duplicate, use 'rocprim::equal_to' in functional.hpp instead.")]] equality
 {
     /// \brief Invocation operator
     template<class T>
@@ -58,7 +52,8 @@ struct equality
 };
 
 /// \brief Functor that tests for inequality.
-struct inequality
+struct [[deprecated(
+    "Duplicate, use 'rocprim::not_equal_to' in functional.hpp instead.")]] inequality
 {
     /// \brief Invocation operator
     template<class T>
@@ -93,7 +88,7 @@ struct inequality_wrapper
 };
 
 /// \brief Functor that returns the sum of its arguments.
-struct sum
+struct [[deprecated("Duplicate, use 'rocprim::plus' in functional.hpp instead.")]] sum
 {
     /// \brief Invocation operator
     template<class T>
@@ -105,7 +100,7 @@ struct sum
 };
 
 /// \brief Functor that returns the maximum of its arguments.
-struct max
+struct [[deprecated("Duplicate, use 'rocprim::max' in functional.hpp instead.")]] max
 {
     /// \brief Invocation operator
     template<class T>
@@ -117,7 +112,7 @@ struct max
 };
 
 /// \brief Functor that returns the minimum of its arguments.
-struct min
+struct [[deprecated("Duplicate, use 'rocprim::min' in functional.hpp instead.")]] min
 {
     /// \brief Invocation operator
     template<class T>
@@ -181,69 +176,6 @@ struct arg_min
 
 /// @}
 // end group thread_operators
-
-namespace detail
-{
-
-// CUB uses value_type of OutputIteratorT (if not void) as a type of intermediate results in scan and reduce,
-// for example:
-//
-// /// The output value type
-// typedef typename If<(Equals<typename std::iterator_traits<OutputIteratorT>::value_type, void>::VALUE),  // OutputT =  (if output iterator's value type is void) ?
-//     typename std::iterator_traits<InputIteratorT>::value_type,                                          // ... then the input iterator's value type,
-//     typename std::iterator_traits<OutputIteratorT>::value_type>::Type OutputT;                          // ... else the output iterator's value type
-//
-// rocPRIM (as well as Thrust) uses result type of BinaryFunction instead (if not void):
-//
-// using input_type = typename std::iterator_traits<InputIterator>::value_type;
-// using result_type = typename ::rocprim::match_result_type<
-//     input_type, BinaryFunction
-// >::type;
-//
-// For short -> float using Sum()
-// CUB:     float Sum(float, float)
-// rocPRIM: short Sum(short, short)
-//
-// This wrapper allows to have compatibility with CUB in hipCUB.
-template<
-    class InputIteratorT,
-    class OutputIteratorT,
-    class BinaryFunction
->
-struct convert_result_type_wrapper
-{
-    using input_type = typename std::iterator_traits<InputIteratorT>::value_type;
-    using output_type = typename std::iterator_traits<OutputIteratorT>::value_type;
-    using result_type =
-        typename std::conditional<
-            std::is_void<output_type>::value, input_type, output_type
-        >::type;
-
-    convert_result_type_wrapper(BinaryFunction op) : op(op) {}
-
-    template<class T>
-    ROCPRIM_HOST_DEVICE inline
-    constexpr result_type operator()(const T &a, const T &b) const
-    {
-        return static_cast<result_type>(op(a, b));
-    }
-
-    BinaryFunction op;
-};
-
-template<
-    class InputIteratorT,
-    class OutputIteratorT,
-    class BinaryFunction
->
-inline
-convert_result_type_wrapper<InputIteratorT, OutputIteratorT, BinaryFunction>
-convert_result_type(BinaryFunction op)
-{
-    return convert_result_type_wrapper<InputIteratorT, OutputIteratorT, BinaryFunction>(op);
-}
-
-} // end detail namespace
 
 END_ROCPRIM_NAMESPACE